+++ include/gaming.h
@@ -2,6 +2,6 @@
 #define __MY_GAMING__
 
 char** InitMap(int size);
-double RayCast(double posX, double posY, double angle, char** map, int size);
+float RayCast(float posX, float posY, float angle, char** map, int size);
 
 #endif
